Pdf_chat_app/requirements.txt
langchain
streamlit
streamlit-extras
faiss-cpu
PyPDF2
python-dotenv
openai
tiktoken
tensorflow
sentence-transformers
InstructorEmbedding